Run tools tests on VR to increase code cov
